Do not need translate when source crop is not (0, 0)
In Android, when the source crop is not (0, 0), the surface damage
is already translated to global display coordinate, no translation
is needed.
When the source crop is (0, 0), no scaling is needed, just transofrm
the coordiate according to the rotation scenario.

Change-Id: I139567a9e59e44a95343adeb53e53b1d5969f581
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-72457
Tests: The HVAC, independent rotation and Simple Podcatcher work well
Signed-off-by: HeYue <yue.he@intel.com>
Reviewed-on: https://android.intel.com:443/654052